+++ astropy/io/ascii/connect.py
@@ -21,15 +21,14 @@ def read_asciitable(filename, **kwargs):
     from .ui import read
     return read(filename, **kwargs)
 
-io_registry.register_reader('ascii', Table, read_asciitable, update_docs=False)
+io_registry.register_reader('ascii', Table, read_asciitable)
 
 
 def write_asciitable(table, filename, **kwargs):
     from .ui import write
     return write(table, filename, **kwargs)
 
-io_registry.register_writer('ascii', Table, write_asciitable,
-                            update_docs=False)
+io_registry.register_writer('ascii', Table, write_asciitable)
 
 
 def io_read(format, filename, **kwargs):
@@ -89,6 +88,6 @@ def write_csv(table, filename, **kwargs):
 
 csv_identify = functools.partial(io_identify, '.csv')
 
-io_registry.register_reader('csv', Table, read_csv, update_docs=False)
-io_registry.register_writer('csv', Table, write_csv, update_docs=False)
+io_registry.register_reader('csv', Table, read_csv)
+io_registry.register_writer('csv', Table, write_csv)
 io_registry.register_identifier('csv', Table, csv_identify)